Move each class from media_player.py into its own file, move consts and funcs into their own appropriate importable files
